Misc Tab Page Changes (Also LINQ Methods) and development information
-Added the functionality to skip the intro video (Not completely done as
the methods now leave extra blankspaces in the ini files which is just
messy, I will fix this next commit)
-Added the foundation for a method that uses LINQ to search for stuff in
an array and delete it out of, I will use this when deleting lines in
files, as it is fast and it will leave no blank space)

Exams are coming up so development on this piece of software is slowing
down, but it will pick back up after the Christmas holidays, in January.
The software is coming along nicely!